+++ src/Kupo/Data/Ogmios.hs
@@ -421,10 +421,11 @@ decodePartialTransaction = Json.withObject "PartialTransaction" $ \o -> do
     case inputSource of
         Just ("collaterals" :: Text) -> do
             inputs <- traverse decodeInput =<< (o .: "body" >>= (.: "collaterals"))
+            outs <- traverse decodeOutput =<< (o .: "body" >>= (.: "collateralReturn"))
             pure PartialTransaction
                 { id
                 , inputs
-                , outputs = []
+                , outputs = withReferences id (maybeToList outs)
                 , datums
                 , scripts
                 , metadata